Refactor common functionality for testing
